requirements.txt
django==4.2.14
PyMySQL==1.1.1
python-decouple==3.8
openai>=1.0.0
httpx>=0.25.0
pandas==2.2.2
openpyxl==3.1.5
django-htmx==1.17.2
python-docx==1.1.2
widget-tweaks==1.5.0
gunicorn
cryptography
APScheduler==3.10.4
pytz